requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
asgiref==3.8.1
    # via django
django==5.2.3
    # via
    #   -r requirements.in
    #   djangorestframework
    #   djangorestframework-simplejwt
djangorestframework==3.16.0
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
djangorestframework-simplejwt==5.5.0
    # via -r requirements.in
psycopg[binary]==3.2.9
    # via -r requirements.in
psycopg-binary==3.2.9
    # via psycopg
pyjwt==2.9.0
    # via djangorestframework-simplejwt
python-dotenv==1.1.0
    # via -r requirements.in
sqlparse==0.5.3
    # via django
typing-extensions==4.14.0
    # via psycopg
tzdata==2025.2
    # via
    #   django
    #   psycopg
